+++ mg_custom_nodes/KLL535_ComfyUI_Simple_Qwen3-VL-gguf_20260831/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Nightly
+- Add dynamic image, audio, video input
+- Add "user_prompt_template" input
+- Add "bypass" input
+- Add "Fix Batch Images" node
+- Add "remove_thinking" config
 - Fix bugs
 - Added new configurator `🌐 System Preset Config (Advanced)`
 - Fix default value